move velocity field class to advection field file
