+++ src/festim/helpers.py
@@ -1,9 +1,6 @@
 from collections.abc import Callable
 from typing import Optional
 
-import basix
-import basix.ufl
-
 import dolfinx
 import numpy as np
 import ufl
@@ -280,88 +277,6 @@ class Value:
             elif isinstance(self.fenics_object, fem.Function):
                 if self.fenics_interpolation_expression is not None:
                     self.fenics_object.interpolate(self.fenics_interpolation_expression)
-
-
-class VelocityField(Value):
-    """
-    A class to handle input values of velocity fields from users and convert them to a
-    relevent fenics object
-
-    Args:
-        input_value: The value of the user input
-
-    Attributes:
-        input_value : The value of the user input
-        fenics_interpolation_expression : The expression of the user input that is used
-            to update the `fenics_object`
-        fenics_object : The value of the user input in fenics format
-        explicit_time_dependent : True if the user input value is explicitly time
-            dependent
-        temperature_dependent : True if the user input value is temperature dependent
-        vector_function_space: the vector function space of the fenics object
-    """
-
-    input_value: fem.Function | Callable
-
-    fenics_object: fem.Function
-    explicit_time_dependent: bool
-    temperature_dependent: bool
-    vector_function_space: fem.FunctionSpace
-
-    def convert_input_value(
-        self,
-        function_space: dolfinx.fem.FunctionSpace,
-        t: Optional[fem.Constant] = None,
-    ):
-        """Converts a user given value to a relevent fenics object
-
-        Args:
-            function_space: the function space of the fenics object
-            t: the time, optional
-        """
-
-        if isinstance(self.input_value, fem.Function):
-            vel = self.input_value
-
-        elif callable(self.input_value):
-            # if callable function has args other than time, t, raise Typer Error
-            args = self.input_value.__code__.co_varnames
-            if args != ("t",):
-                raise TypeError(
-                    "velocity function can only be a function of time arg t"
-                )
-
-            vel = self.input_value(t)
-
-            # if function does not return an fem.Fucntion, raise Typer Error
-            if not isinstance(vel, fem.Function):
-                raise ValueError(
-                    "A time dependent advection field should return an fem.Function"
-                    f", not a {type(vel)}"
-                )
-
-        # create vector function space and function
-        v_cg = basix.ufl.element(
-            "Lagrange",
-            function_space.mesh.topology.cell_name(),
-            1,
-            shape=(function_space.mesh.topology.dim,),
-        )
-        self.vector_function_space = dolfinx.fem.functionspace(
-            function_space.mesh, v_cg
-        )
-        self.fenics_object = fem.Function(self.vector_function_space)
-
-        # interpolate input value into fenics object function
-        nmm_interpolate(self.fenics_object, vel)
-
-    def update(self, t: fem.Constant):
-        """Updates the velocity field
-
-        Args:
-            t: the time
-        """
-        nmm_interpolate(self.fenics_object, self.input_value(t=float(t)))
 
 
 # Check the version of dolfinx
